+++ src/index.ts
@@ -1,5 +1,5 @@
 export const packageName = '@standhigher/charts';
-export const packageVersion = '1.0.0';
+export const packageVersion = '1.1.0';
 
 export * from './formatters';
 export * from './presets';